게시글 수정 null point 에러 수정

+++ src/main/java/com/one/whospet/service/board/impl/BoardServiceImpl.java
@@ -314,20 +314,28 @@ public class BoardServiceImpl implements BoardService {
 		
 		//유저가 삭제선택한 이미지 정보가 있는지 체크
 		int count = 1;
-		int files = fileRequest.getParameterValues("fileInfos").length; //선택된 파일들 갯수
-		while(count <= files) {
-			String fileInfo = fileRequest.getParameter("fileInfo" + count);
-			if(fileInfo != null && !fileInfo.equals("")) {
-				boardDao.deleteCheckFile(fileInfo); // db에서 선택된파일지움			
-				String path = context.getRealPath("upload");	
-				//현재 게시판에 존재하는 파일객체를 만듬
-				File file = new File(path + "\\" + fileInfo);		
-				if(file.exists()) { // 파일이 존재하면
-					file.delete(); // 파일 삭제	
-				}	
-			}
-			count++; //다음체크파일 확인하기 위해 카운트 변수 더해줌
+		Integer files = 0;
+		try {
+			files = fileRequest.getParameterValues("fileInfos").length; //선택된 파일들 갯수
+		}catch (Exception e) {
+			files = 0;
 		}
+		if(files != null && files != 0) {
+			while(count <= files) {
+				String fileInfo = fileRequest.getParameter("fileInfo" + count);
+				if(fileInfo != null && !fileInfo.equals("")) {
+					boardDao.deleteCheckFile(fileInfo); // db에서 선택된파일지움			
+					String path = context.getRealPath("upload");	
+					//현재 게시판에 존재하는 파일객체를 만듬
+					File file = new File(path + "\\" + fileInfo);		
+					if(file.exists()) { // 파일이 존재하면
+						file.delete(); // 파일 삭제	
+					}	
+				}
+				count++; //다음체크파일 확인하기 위해 카운트 변수 더해줌
+			}
+		}
+		
 		
 			
 		//첨부파일이 null이거나 없으면 게시글 파일이 없으면 작성 끝
